Update ESP_DataImporter_NWRFC.py
Changed a zero value to "Constants.UNDEFINED". This allows for an missing value to be put in the time series container so the user know data is missing. Having a zero value might be misleading as some locations might get really low where a zero might be an acceptable value.